Move pathEndRuntime methods into path_end_runtime.go. Move PathProcessor private methods to path_processor_internal.go
