+++ relayer/processor/path_processor_internal.go
@@ -5,212 +5,493 @@ import (
 	"errors"
 	"fmt"
 
+	"golang.org/x/sync/errgroup"
 	"github.com/cosmos/relayer/v2/relayer/provider"
 	"go.uber.org/zap"
 )
 
-// shouldSendPacketMessage determines if the packet flow message should be sent now.
-// It will also determine if the message needs to be given up on entirely and remove retention if so.
-func (pathEnd *pathEndRuntime) shouldSendPacketMessage(message packetIBCMessage, counterparty *pathEndRuntime) bool {
-	if !pathEnd.channelStateCache[message.channelKey] {
-		// channel is not open, do not send
-		return false
+// assembleIBCMessage constructs the applicable IBC message using the requested function.
+// These functions may do things like make queries in order to assemble a complete IBC message.
+func (pp *PathProcessor) assemblePacketIBCMessage(
+	ctx context.Context,
+	src, dst *pathEndRuntime,
+	partialMessage packetIBCMessage,
+	assembleMessage func(ctx context.Context, msgRecvPacket provider.RelayerMessage, signer string, latest provider.LatestBlock) (provider.RelayerMessage, error),
+) (provider.RelayerMessage, error) {
+	signer, err := dst.chainProvider.Address()
+	if err != nil {
+		return nil, fmt.Errorf("error getting signer address for {%s}: %w", dst.info.ChainID, err)
 	}
-	msgSendCache, ok := pathEnd.packetSendCache[message.channelKey]
-	if !ok {
-		// in progress cache does not exist for this channel, so can send.
-		return true
+	assembled, err := assembleMessage(ctx, partialMessage.message, signer, src.latestBlock)
+	if err != nil {
+		return nil, fmt.Errorf("error assembling %s for {%s}: %w", partialMessage.action, dst.info.ChainID, err)
 	}
-	channelSendCache, ok := msgSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send
-		return true
+
+	return assembled, nil
+}
+
+func (pp *PathProcessor) getUnrelayedPacketsAndAcksAndToDelete(ctx context.Context, pathEndPacketFlowMessages pathEndPacketFlowMessages) pathEndPacketFlowResponse {
+	res := pathEndPacketFlowResponse{
+		ToDeleteSrc: make(map[string][]uint64),
+		ToDeleteDst: make(map[string][]uint64),
 	}
-	inProgress, ok := channelSendCache[message.sequence]
-	if !ok {
-		// in progress cache does not exist for this sequence, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this packet flow message
-		// remove all retention of this connection handshake in pathEnd.messagesCache.PacketFlow and counterparty
-		toDelete := make(map[string][]uint64)
-		toDeleteCounterparty := make(map[string][]uint64)
-		switch message.action {
-		case MsgRecvPacket:
-			toDelete[MsgRecvPacket] = []uint64{message.sequence}
-			toDeleteCounterparty[MsgTransfer] = []uint64{message.sequence}
-		case MsgAcknowledgement, MsgTimeout, MsgTimeoutOnClose:
-			toDelete[message.action] = []uint64{message.sequence}
-			toDeleteCounterparty[MsgRecvPacket] = []uint64{message.sequence}
-			toDelete[MsgTransfer] = []uint64{message.sequence}
+
+MsgTransferLoop:
+	for transferSeq, msgTransfer := range pathEndPacketFlowMessages.SrcMsgTransfer {
+		for ackSeq := range pathEndPacketFlowMessages.SrcMsgAcknowledgement {
+			if transferSeq == ackSeq {
+				// we have an ack for this packet, so packet flow is complete
+				// remove all retention of this sequence number
+				res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], transferSeq)
+				res.ToDeleteDst[MsgRecvPacket] = append(res.ToDeleteDst[MsgRecvPacket], transferSeq)
+				res.ToDeleteSrc[MsgAcknowledgement] = append(res.ToDeleteSrc[MsgAcknowledgement], transferSeq)
+				continue MsgTransferLoop
+			}
 		}
-		// delete in progress send for this specific message
-		pathEnd.packetSendCache[message.channelKey].deleteCachedMessages(map[string][]uint64{message.action: []uint64{message.sequence}})
-		// delete all packet flow retention history for this sequence
-		pathEnd.messageCache.PacketFlow[message.channelKey].DeleteCachedMessages(toDelete)
-		counterparty.messageCache.PacketFlow[message.channelKey].DeleteCachedMessages(toDeleteCounterparty)
-		return false
-	}
-
-	return true
-}
-
-// shouldSendConnectionMessage determines if the connection handshake message should be sent now.
-// It will also determine if the message needs to be given up on entirely and remove retention if so.
-func (pathEnd *pathEndRuntime) shouldSendConnectionMessage(message connectionIBCMessage, counterparty *pathEndRuntime) bool {
-	msgSendCache, ok := pathEnd.connectionSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send.
-		return true
-	}
-	inProgress, ok := msgSendCache[message.connectionKey]
-	if !ok {
-		// in progress cache does not exist for this connection, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this connection handshake message
-		// remove all retention of this connection handshake in pathEnd.messagesCache.ConnectionHandshake and counterparty
-		toDelete := make(map[string][]ConnectionKey)
-		toDeleteCounterparty := make(map[string][]ConnectionKey)
-		switch message.action {
-		case MsgConnectionOpenTry:
-			toDeleteCounterparty[MsgConnectionOpenInit] = []ConnectionKey{message.connectionKey}
-		case MsgConnectionOpenAck:
-			toDeleteCounterparty[MsgConnectionOpenTry] = []ConnectionKey{message.connectionKey}
-			toDelete[MsgConnectionOpenInit] = []ConnectionKey{message.connectionKey}
-		case MsgConnectionOpenConfirm:
-			toDeleteCounterparty[MsgConnectionOpenAck] = []ConnectionKey{message.connectionKey}
-			toDelete[MsgConnectionOpenTry] = []ConnectionKey{message.connectionKey}
-			toDeleteCounterparty[MsgConnectionOpenInit] = []ConnectionKey{message.connectionKey}
+		for timeoutSeq := range pathEndPacketFlowMessages.SrcMsgTimeout {
+			if transferSeq == timeoutSeq {
+				// we have a timeout for this packet, so packet flow is complete
+				// remove all retention of this sequence number
+				res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], transferSeq)
+				res.ToDeleteSrc[MsgTimeout] = append(res.ToDeleteSrc[MsgTimeout], transferSeq)
+				continue MsgTransferLoop
+			}
 		}
-		// delete in progress send for this specific message
-		pathEnd.connectionSendCache.deleteCachedMessages(map[string][]ConnectionKey{message.action: []ConnectionKey{message.connectionKey}})
-		// delete all connection handshake retention history for this connection
-		pathEnd.messageCache.ConnectionHandshake.DeleteCachedMessages(toDelete)
-		counterparty.messageCache.ConnectionHandshake.DeleteCachedMessages(toDeleteCounterparty)
-		return false
-	}
-
-	return true
-}
-
-// shouldSendConnectionMessage determines if the channel handshake message should be sent now.
-// It will also determine if the message needs to be given up on entirely and remove retention if so.
-func (pathEnd *pathEndRuntime) shouldSendChannelMessage(message channelIBCMessage, counterparty *pathEndRuntime) bool {
-	msgSendCache, ok := pathEnd.channelSendCache[message.action]
-	if !ok {
-		// in progress cache does not exist for this action, so can send.
-		return true
-	}
-	inProgress, ok := msgSendCache[message.channelKey]
-	if !ok {
-		// in progress cache does not exist for this channel, so can send.
-		return true
-	}
-	if pathEnd.latestBlock.Height-inProgress.sendHeight < blocksToRetryPacketAfter {
-		// this message was sent less than blocksToRetryPacketAfter ago, do not send again yet.
-		return false
-	}
-	if inProgress.retryCount == maxMessageSendRetries {
-		// giving up on sending this channel handshake message
-		// remove all retention of this connection handshake in pathEnd.messagesCache.ConnectionHandshake and counterparty
-		toDelete := make(map[string][]ChannelKey)
-		toDeleteCounterparty := make(map[string][]ChannelKey)
-		switch message.action {
-		case MsgChannelOpenTry:
-			toDeleteCounterparty[MsgChannelOpenInit] = []ChannelKey{message.channelKey}
-		case MsgChannelOpenAck:
-			toDeleteCounterparty[MsgChannelOpenTry] = []ChannelKey{message.channelKey}
-			toDelete[MsgChannelOpenInit] = []ChannelKey{message.channelKey}
-		case MsgChannelOpenConfirm:
-			toDeleteCounterparty[MsgChannelOpenAck] = []ChannelKey{message.channelKey}
-			toDelete[MsgChannelOpenTry] = []ChannelKey{message.channelKey}
-			toDeleteCounterparty[MsgChannelOpenInit] = []ChannelKey{message.channelKey}
+		for timeoutOnCloseSeq := range pathEndPacketFlowMessages.SrcMsgTimeoutOnClose {
+			if transferSeq == timeoutOnCloseSeq {
+				// we have a timeout for this packet, so packet flow is complete
+				// remove all retention of this sequence number
+				res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], transferSeq)
+				res.ToDeleteSrc[MsgTimeoutOnClose] = append(res.ToDeleteSrc[MsgTimeoutOnClose], transferSeq)
+				continue MsgTransferLoop
+			}
 		}
-		// delete in progress send for this specific message
-		pathEnd.channelSendCache.deleteCachedMessages(map[string][]ChannelKey{message.action: []ChannelKey{message.channelKey}})
-		// delete all connection handshake retention history for this channel
-		pathEnd.messageCache.ChannelHandshake.DeleteCachedMessages(toDelete)
-		counterparty.messageCache.ChannelHandshake.DeleteCachedMessages(toDeleteCounterparty)
-		return false
+		for msgRecvSeq, msgAcknowledgement := range pathEndPacketFlowMessages.DstMsgRecvPacket {
+			if transferSeq == msgRecvSeq {
+				// msg is received by dst chain, but no ack yet. Need to relay ack from dst to src!
+				ackMsg := packetIBCMessage{
+					channelKey: pathEndPacketFlowMessages.ChannelKey,
+					action:     MsgAcknowledgement,
+					sequence:   transferSeq,
+					message:    msgAcknowledgement,
+				}
+				if pathEndPacketFlowMessages.Src.shouldSendPacketMessage(ackMsg, pathEndPacketFlowMessages.Dst) {
+					res.SrcMessages = append(res.SrcMessages, ackMsg)
+				}
+				continue MsgTransferLoop
+			}
+		}
+		// Packet is not yet relayed! need to relay either MsgRecvPacket from src to dst, or MsgTimeout/MsgTimeoutOnClose from dst to src
+		if err := pathEndPacketFlowMessages.Dst.chainProvider.ValidatePacket(msgTransfer, pathEndPacketFlowMessages.Dst.latestBlock); err != nil {
+			var timeoutHeightErr *provider.TimeoutHeightError
+			var timeoutTimestampErr *provider.TimeoutTimestampError
+			var timeoutOnCloseErr *provider.TimeoutOnCloseError
+
+			switch {
+			case errors.As(err, &timeoutHeightErr) || errors.As(err, &timeoutTimestampErr):
+				timeoutMsg := packetIBCMessage{
+					channelKey: pathEndPacketFlowMessages.ChannelKey,
+					action:     MsgTimeout,
+					sequence:   transferSeq,
+					message:    msgTransfer,
+				}
+				if pathEndPacketFlowMessages.Src.shouldSendPacketMessage(timeoutMsg, pathEndPacketFlowMessages.Dst) {
+					res.SrcMessages = append(res.SrcMessages, timeoutMsg)
+				}
+			case errors.As(err, &timeoutOnCloseErr):
+				timeoutOnCloseMsg := packetIBCMessage{
+					channelKey: pathEndPacketFlowMessages.ChannelKey,
+					action:     MsgTimeoutOnClose,
+					sequence:   transferSeq,
+					message:    msgTransfer,
+				}
+				if pathEndPacketFlowMessages.Src.shouldSendPacketMessage(timeoutOnCloseMsg, pathEndPacketFlowMessages.Dst) {
+					res.SrcMessages = append(res.SrcMessages, timeoutOnCloseMsg)
+				}
+			default:
+				pp.log.Error("Packet is invalid",
+					zap.String("chain_id", pathEndPacketFlowMessages.Src.info.ChainID),
+					zap.Error(err),
+				)
+			}
+			continue MsgTransferLoop
+		}
+		recvPacketMsg := packetIBCMessage{
+			channelKey: pathEndPacketFlowMessages.ChannelKey.Counterparty(),
+			action:     MsgRecvPacket,
+			sequence:   transferSeq,
+			message:    msgTransfer,
+		}
+		if pathEndPacketFlowMessages.Dst.shouldSendPacketMessage(recvPacketMsg, pathEndPacketFlowMessages.Src) {
+			res.DstMessages = append(res.DstMessages, recvPacketMsg)
+		}
 	}
 
-	return true
+	// now iterate through packet-flow-complete messages and remove any leftover messages if the MsgTransfer or MsgRecvPacket was in a previous block that we did not query
+	for ackSeq := range pathEndPacketFlowMessages.SrcMsgAcknowledgement {
+		res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], ackSeq)
+		res.ToDeleteDst[MsgRecvPacket] = append(res.ToDeleteDst[MsgRecvPacket], ackSeq)
+		res.ToDeleteSrc[MsgAcknowledgement] = append(res.ToDeleteSrc[MsgAcknowledgement], ackSeq)
+	}
+	for timeoutSeq := range pathEndPacketFlowMessages.SrcMsgTimeout {
+		res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], timeoutSeq)
+		res.ToDeleteSrc[MsgTimeout] = append(res.ToDeleteSrc[MsgTimeout], timeoutSeq)
+	}
+	for timeoutOnCloseSeq := range pathEndPacketFlowMessages.SrcMsgTimeoutOnClose {
+		res.ToDeleteSrc[MsgTransfer] = append(res.ToDeleteSrc[MsgTransfer], timeoutOnCloseSeq)
+		res.ToDeleteSrc[MsgTimeoutOnClose] = append(res.ToDeleteSrc[MsgTimeoutOnClose], timeoutOnCloseSeq)
+	}
+	return res
 }
 
-func (pathEnd *pathEndRuntime) trackSentPacketMessage(message packetIBCMessage) {
-	msgSendCache, ok := pathEnd.packetSendCache[message.channelKey]
+func (pp *PathProcessor) getUnrelayedConnectionHandshakeMessagesAndToDelete(pathEndConnectionHandshakeMessages pathEndConnectionHandshakeMessages) pathEndConnectionHandshakeResponse {
+	res := pathEndConnectionHandshakeResponse{
+		ToDeleteSrc: make(map[string][]ConnectionKey),
+		ToDeleteDst: make(map[string][]ConnectionKey),
+	}
+
+ConnectionHandshakeLoop:
+	for openInitKey, openInitMsg := range pathEndConnectionHandshakeMessages.SrcMsgConnectionOpenInit {
+		var foundOpenTry provider.RelayerMessage
+		for openTryKey, openTryMsg := range pathEndConnectionHandshakeMessages.DstMsgConnectionOpenTry {
+			if openInitKey == openTryKey {
+				foundOpenTry = openTryMsg
+				break
+			}
+		}
+		if foundOpenTry == nil {
+			// need to send an open try to dst
+			msgOpenTry := connectionIBCMessage{
+				action:        MsgConnectionOpenTry,
+				connectionKey: openInitKey.Counterparty(),
+				message:       openInitMsg,
+			}
+			if pathEndConnectionHandshakeMessages.Dst.shouldSendConnectionMessage(msgOpenTry, pathEndConnectionHandshakeMessages.Src) {
+				res.DstMessages = append(res.DstMessages, msgOpenTry)
+			}
+			continue ConnectionHandshakeLoop
+		}
+		var foundOpenAck provider.RelayerMessage
+		for openAckKey, openAckMsg := range pathEndConnectionHandshakeMessages.SrcMsgConnectionOpenAck {
+			if openInitKey == openAckKey {
+				foundOpenAck = openAckMsg
+				break
+			}
+		}
+		if foundOpenAck == nil {
+			// need to send an open ack to src
+			msgOpenAck := connectionIBCMessage{
+				action:        MsgConnectionOpenAck,
+				connectionKey: openInitKey,
+				message:       foundOpenTry,
+			}
+			if pathEndConnectionHandshakeMessages.Src.shouldSendConnectionMessage(msgOpenAck, pathEndConnectionHandshakeMessages.Dst) {
+				res.SrcMessages = append(res.SrcMessages, msgOpenAck)
+			}
+			continue ConnectionHandshakeLoop
+		}
+		var foundOpenConfirm provider.RelayerMessage
+		for openConfirmKey, openConfirmMsg := range pathEndConnectionHandshakeMessages.DstMsgConnectionOpenConfirm {
+			if openInitKey == openConfirmKey {
+				foundOpenConfirm = openConfirmMsg
+				break
+			}
+		}
+		if foundOpenConfirm == nil {
+			// need to send an open confirm to dst
+			msgOpenConfirm := connectionIBCMessage{
+				action:        MsgConnectionOpenConfirm,
+				connectionKey: openInitKey.Counterparty(),
+				message:       foundOpenAck,
+			}
+			if pathEndConnectionHandshakeMessages.Dst.shouldSendConnectionMessage(msgOpenConfirm, pathEndConnectionHandshakeMessages.Src) {
+				res.DstMessages = append(res.DstMessages, msgOpenConfirm)
+			}
+			continue ConnectionHandshakeLoop
+		}
+		// handshake is complete for this connection, remove all retention.
+		res.ToDeleteSrc[MsgConnectionOpenInit] = append(res.ToDeleteSrc[MsgConnectionOpenInit], openInitKey)
+		res.ToDeleteDst[MsgConnectionOpenTry] = append(res.ToDeleteDst[MsgConnectionOpenTry], openInitKey)
+		res.ToDeleteSrc[MsgConnectionOpenAck] = append(res.ToDeleteSrc[MsgConnectionOpenAck], openInitKey)
+		res.ToDeleteDst[MsgConnectionOpenConfirm] = append(res.ToDeleteDst[MsgConnectionOpenConfirm], openInitKey)
+	}
+
+	// now iterate through connection-handshake-complete messages and remove any leftover messages
+	for openConfirmKey := range pathEndConnectionHandshakeMessages.DstMsgConnectionOpenConfirm {
+		res.ToDeleteSrc[MsgConnectionOpenInit] = append(res.ToDeleteSrc[MsgConnectionOpenInit], openConfirmKey)
+		res.ToDeleteDst[MsgConnectionOpenTry] = append(res.ToDeleteDst[MsgConnectionOpenTry], openConfirmKey)
+		res.ToDeleteSrc[MsgConnectionOpenAck] = append(res.ToDeleteSrc[MsgConnectionOpenAck], openConfirmKey)
+		res.ToDeleteDst[MsgConnectionOpenConfirm] = append(res.ToDeleteDst[MsgConnectionOpenConfirm], openConfirmKey)
+	}
+	return res
+}
+
+func (pp *PathProcessor) getUnrelayedChannelHandshakeMessagesAndToDelete(pathEndChannelHandshakeMessages pathEndChannelHandshakeMessages) pathEndChannelHandshakeResponse {
+	res := pathEndChannelHandshakeResponse{
+		ToDeleteSrc: make(map[string][]ChannelKey),
+		ToDeleteDst: make(map[string][]ChannelKey),
+	}
+
+ChannelHandshakeLoop:
+	for openInitKey, openInitMsg := range pathEndChannelHandshakeMessages.SrcMsgChannelOpenInit {
+		var foundOpenTry provider.RelayerMessage
+		for openTryKey, openTryMsg := range pathEndChannelHandshakeMessages.DstMsgChannelOpenTry {
+			if openInitKey == openTryKey {
+				foundOpenTry = openTryMsg
+				break
+			}
+		}
+		if foundOpenTry == nil {
+			// need to send an open try to dst
+			msgOpenTry := channelIBCMessage{
+				action:     MsgChannelOpenTry,
+				channelKey: openInitKey.Counterparty(),
+				message:    openInitMsg,
+			}
+			if pathEndChannelHandshakeMessages.Dst.shouldSendChannelMessage(msgOpenTry, pathEndChannelHandshakeMessages.Src) {
+				res.DstMessages = append(res.DstMessages, msgOpenTry)
+			}
+			continue ChannelHandshakeLoop
+		}
+		var foundOpenAck provider.RelayerMessage
+		for openAckKey, openAckMsg := range pathEndChannelHandshakeMessages.SrcMsgChannelOpenAck {
+			if openInitKey == openAckKey {
+				foundOpenAck = openAckMsg
+				break
+			}
+		}
+		if foundOpenAck == nil {
+			// need to send an open ack to src
+			msgOpenAck := channelIBCMessage{
+				action:     MsgChannelOpenAck,
+				channelKey: openInitKey,
+				message:    foundOpenTry,
+			}
+			if pathEndChannelHandshakeMessages.Src.shouldSendChannelMessage(msgOpenAck, pathEndChannelHandshakeMessages.Dst) {
+				res.SrcMessages = append(res.SrcMessages, msgOpenAck)
+			}
+			continue ChannelHandshakeLoop
+		}
+		var foundOpenConfirm provider.RelayerMessage
+		for openConfirmKey, openConfirmMsg := range pathEndChannelHandshakeMessages.DstMsgChannelOpenConfirm {
+			if openInitKey == openConfirmKey {
+				foundOpenConfirm = openConfirmMsg
+				break
+			}
+		}
+		if foundOpenConfirm == nil {
+			// need to send an open confirm to dst
+			msgOpenConfirm := channelIBCMessage{
+				action:     MsgChannelOpenConfirm,
+				channelKey: openInitKey.Counterparty(),
+				message:    foundOpenAck,
+			}
+			if pathEndChannelHandshakeMessages.Dst.shouldSendChannelMessage(msgOpenConfirm, pathEndChannelHandshakeMessages.Src) {
+				res.DstMessages = append(res.DstMessages, msgOpenConfirm)
+			}
+			continue ChannelHandshakeLoop
+		}
+		// handshake is complete for this channel, remove all retention.
+		res.ToDeleteSrc[MsgChannelOpenInit] = append(res.ToDeleteSrc[MsgChannelOpenInit], openInitKey)
+		res.ToDeleteDst[MsgChannelOpenTry] = append(res.ToDeleteDst[MsgChannelOpenTry], openInitKey)
+		res.ToDeleteSrc[MsgChannelOpenAck] = append(res.ToDeleteSrc[MsgChannelOpenAck], openInitKey)
+		res.ToDeleteDst[MsgChannelOpenConfirm] = append(res.ToDeleteDst[MsgChannelOpenConfirm], openInitKey)
+	}
+
+	// now iterate through channel-handshake-complete messages and remove any leftover messages
+	for openConfirmKey := range pathEndChannelHandshakeMessages.DstMsgChannelOpenConfirm {
+		res.ToDeleteSrc[MsgChannelOpenInit] = append(res.ToDeleteSrc[MsgChannelOpenInit], openConfirmKey)
+		res.ToDeleteDst[MsgChannelOpenTry] = append(res.ToDeleteDst[MsgChannelOpenTry], openConfirmKey)
+		res.ToDeleteSrc[MsgChannelOpenAck] = append(res.ToDeleteSrc[MsgChannelOpenAck], openConfirmKey)
+		res.ToDeleteDst[MsgChannelOpenConfirm] = append(res.ToDeleteDst[MsgChannelOpenConfirm], openConfirmKey)
+	}
+	return res
+}
+
+// assembleMsgUpdateClient uses the ChainProvider from both pathEnds to assemble the client update header
+// from the source and then assemble the update client message in the correct format for the destination.
+func (pp *PathProcessor) assembleMsgUpdateClient(ctx context.Context, src, dst *pathEndRuntime) (provider.RelayerMessage, error) {
+	clientID := dst.info.ClientID
+	clientConsensusHeight := dst.clientState.ConsensusHeight
+	trustedConsensusHeight := dst.clientTrustedState.ClientState.ConsensusHeight
+
+	// If the client state height is not equal to the client trusted state height and the client state height is
+	// the latest block, we cannot send a MsgUpdateClient until another block is observed on the counterparty.
+	// If the client state height is in the past, beyond ibcHeadersToCache, then we need to query for it.
+	if !trustedConsensusHeight.EQ(clientConsensusHeight) {
+		if int64(clientConsensusHeight.RevisionHeight)-int64(trustedConsensusHeight.RevisionHeight) <= ibcHeadersToCache {
+			return nil, fmt.Errorf("observed client trusted height: %d does not equal latest client state height: %d",
+				trustedConsensusHeight.RevisionHeight, clientConsensusHeight.RevisionHeight)
+		}
+		header, err := src.chainProvider.IBCHeaderAtHeight(ctx, int64(clientConsensusHeight.RevisionHeight+1))
+		if err != nil {
+			return nil, fmt.Errorf("error getting IBC header at height: %d for chain_id: %s, %w", clientConsensusHeight.RevisionHeight+1, src.info.ChainID, err)
+		}
+		pp.log.Warn("Had to query for client trusted IBC header",
+			zap.String("chain_id", src.info.ChainID),
+			zap.String("counterparty_chain_id", dst.info.ChainID),
+			zap.String("counterparty_client_id", clientID),
+			zap.Uint64("height", clientConsensusHeight.RevisionHeight+1),
+			zap.Uint64("latest_height", src.latestBlock.Height),
+		)
+		dst.clientTrustedState = provider.ClientTrustedState{
+			ClientState: dst.clientState,
+			IBCHeader:   header,
+		}
+		trustedConsensusHeight = clientConsensusHeight
+	}
+
+	if src.latestHeader.Height() == trustedConsensusHeight.RevisionHeight {
+		return nil, fmt.Errorf("latest header height is equal to the client trusted height: %d, "+
+			"need to wait for next block's header before we can assemble and send a new MsgUpdateClient",
+			trustedConsensusHeight.RevisionHeight)
+	}
+
+	msgUpdateClientHeader, err := src.chainProvider.MsgUpdateClientHeader(src.latestHeader, trustedConsensusHeight, dst.clientTrustedState.IBCHeader)
+	if err != nil {
+		return nil, fmt.Errorf("error assembling new client header: %w", err)
+	}
+
+	msgUpdateClient, err := dst.chainProvider.MsgUpdateClient(clientID, msgUpdateClientHeader)
+	if err != nil {
+		return nil, fmt.Errorf("error assembling MsgUpdateClient: %w", err)
+	}
+
+	return msgUpdateClient, nil
+}
+
+// updateClientTrustedState combines the counterparty chains trusted IBC header
+// with the latest client state, which will be used for constructing MsgUpdateClient messages.
+func (pp *PathProcessor) updateClientTrustedState(src *pathEndRuntime, dst *pathEndRuntime) {
+	if src.clientTrustedState.ClientState.ConsensusHeight.GTE(src.clientState.ConsensusHeight) {
+		// current height already trusted
+		return
+	}
+	// need to assemble new trusted state
+	ibcHeader, ok := dst.ibcHeaderCache[src.clientState.ConsensusHeight.RevisionHeight+1]
 	if !ok {
-		msgSendCache = make(packetChannelMessageCache)
-		pathEnd.packetSendCache[message.channelKey] = msgSendCache
+		pp.log.Warn("No cached IBC header for client trusted height",
+			zap.String("chain_id", src.info.ChainID),
+			zap.String("client_id", src.info.ClientID),
+			zap.Uint64("height", src.clientState.ConsensusHeight.RevisionHeight+1),
+		)
+		return
 	}
-	channelSendCache, ok := msgSendCache[message.action]
-	if !ok {
-		channelSendCache = make(packetMessageSendCache)
-		msgSendCache[message.action] = channelSendCache
-	}
-
-	retryCount := uint64(0)
-
-	if inProgress, ok := channelSendCache[message.sequence]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	channelSendCache[message.sequence] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
+	src.clientTrustedState = provider.ClientTrustedState{
+		ClientState: src.clientState,
+		IBCHeader:   ibcHeader,
 	}
 }
 
-func (pathEnd *pathEndRuntime) trackSentConnectionMessage(message connectionIBCMessage) {
-	msgSendCache, ok := pathEnd.connectionSendCache[message.action]
-	if !ok {
-		msgSendCache = make(connectionKeySendCache)
-		pathEnd.connectionSendCache[message.action] = msgSendCache
+// messages from both pathEnds are needed in order to determine what needs to be relayed for a single pathEnd
+func (pp *PathProcessor) processLatestMessages(ctx context.Context) error {
+	// Update trusted client state for both pathends
+	pp.updateClientTrustedState(pp.pathEnd1, pp.pathEnd2)
+	pp.updateClientTrustedState(pp.pathEnd2, pp.pathEnd1)
+
+	channelPairs := pp.channelPairs()
+
+	pathEnd1ConnectionHandshakeMessages := pathEndConnectionHandshakeMessages{
+		Src:                         pp.pathEnd1,
+		Dst:                         pp.pathEnd2,
+		SrcMsgConnectionOpenInit:    pp.pathEnd1.messageCache.ConnectionHandshake[MsgConnectionOpenInit],
+		DstMsgConnectionOpenTry:     pp.pathEnd2.messageCache.ConnectionHandshake[MsgConnectionOpenTry],
+		SrcMsgConnectionOpenAck:     pp.pathEnd1.messageCache.ConnectionHandshake[MsgConnectionOpenAck],
+		DstMsgConnectionOpenConfirm: pp.pathEnd2.messageCache.ConnectionHandshake[MsgConnectionOpenConfirm],
+	}
+	pathEnd2ConnectionHandshakeMessages := pathEndConnectionHandshakeMessages{
+		Src:                         pp.pathEnd2,
+		Dst:                         pp.pathEnd1,
+		SrcMsgConnectionOpenInit:    pp.pathEnd2.messageCache.ConnectionHandshake[MsgConnectionOpenInit],
+		DstMsgConnectionOpenTry:     pp.pathEnd1.messageCache.ConnectionHandshake[MsgConnectionOpenTry],
+		SrcMsgConnectionOpenAck:     pp.pathEnd2.messageCache.ConnectionHandshake[MsgConnectionOpenAck],
+		DstMsgConnectionOpenConfirm: pp.pathEnd1.messageCache.ConnectionHandshake[MsgConnectionOpenConfirm],
+	}
+	pathEnd1ConnectionHandshakeRes := pp.getUnrelayedConnectionHandshakeMessagesAndToDelete(pathEnd1ConnectionHandshakeMessages)
+	pathEnd2ConnectionHandshakeRes := pp.getUnrelayedConnectionHandshakeMessagesAndToDelete(pathEnd2ConnectionHandshakeMessages)
+
+	pathEnd1ChannelHandshakeMessages := pathEndChannelHandshakeMessages{
+		Src:                      pp.pathEnd1,
+		Dst:                      pp.pathEnd2,
+		SrcMsgChannelOpenInit:    pp.pathEnd1.messageCache.ChannelHandshake[MsgChannelOpenInit],
+		DstMsgChannelOpenTry:     pp.pathEnd2.messageCache.ChannelHandshake[MsgChannelOpenTry],
+		SrcMsgChannelOpenAck:     pp.pathEnd1.messageCache.ChannelHandshake[MsgChannelOpenAck],
+		DstMsgChannelOpenConfirm: pp.pathEnd2.messageCache.ChannelHandshake[MsgChannelOpenConfirm],
+	}
+	pathEnd2ChannelHandshakeMessages := pathEndChannelHandshakeMessages{
+		Src:                      pp.pathEnd2,
+		Dst:                      pp.pathEnd1,
+		SrcMsgChannelOpenInit:    pp.pathEnd2.messageCache.ChannelHandshake[MsgChannelOpenInit],
+		DstMsgChannelOpenTry:     pp.pathEnd1.messageCache.ChannelHandshake[MsgChannelOpenTry],
+		SrcMsgChannelOpenAck:     pp.pathEnd2.messageCache.ChannelHandshake[MsgChannelOpenAck],
+		DstMsgChannelOpenConfirm: pp.pathEnd1.messageCache.ChannelHandshake[MsgChannelOpenConfirm],
+	}
+	pathEnd1ChannelHandshakeRes := pp.getUnrelayedChannelHandshakeMessagesAndToDelete(pathEnd1ChannelHandshakeMessages)
+	pathEnd2ChannelHandshakeRes := pp.getUnrelayedChannelHandshakeMessagesAndToDelete(pathEnd2ChannelHandshakeMessages)
+
+	// process the packet flows for both path ends to determine what needs to be relayed
+	pathEnd1ProcessRes := make([]pathEndPacketFlowResponse, len(channelPairs))
+	pathEnd2ProcessRes := make([]pathEndPacketFlowResponse, len(channelPairs))
+
+	for i, pair := range channelPairs {
+		pathEnd1PacketFlowMessages := pathEndPacketFlowMessages{
+			Src:                   pp.pathEnd1,
+			Dst:                   pp.pathEnd2,
+			ChannelKey:            pair.pathEnd1ChannelKey,
+			SrcMsgTransfer:        pp.pathEnd1.messageCache.PacketFlow[pair.pathEnd1ChannelKey][MsgTransfer],
+			DstMsgRecvPacket:      pp.pathEnd2.messageCache.PacketFlow[pair.pathEnd2ChannelKey][MsgRecvPacket],
+			SrcMsgAcknowledgement: pp.pathEnd1.messageCache.PacketFlow[pair.pathEnd1ChannelKey][MsgAcknowledgement],
+			SrcMsgTimeout:         pp.pathEnd1.messageCache.PacketFlow[pair.pathEnd1ChannelKey][MsgTimeout],
+			SrcMsgTimeoutOnClose:  pp.pathEnd1.messageCache.PacketFlow[pair.pathEnd1ChannelKey][MsgTimeoutOnClose],
+		}
+		pathEnd2PacketFlowMessages := pathEndPacketFlowMessages{
+			Src:                   pp.pathEnd2,
+			Dst:                   pp.pathEnd1,
+			ChannelKey:            pair.pathEnd2ChannelKey,
+			SrcMsgTransfer:        pp.pathEnd2.messageCache.PacketFlow[pair.pathEnd2ChannelKey][MsgTransfer],
+			DstMsgRecvPacket:      pp.pathEnd1.messageCache.PacketFlow[pair.pathEnd1ChannelKey][MsgRecvPacket],
+			SrcMsgAcknowledgement: pp.pathEnd2.messageCache.PacketFlow[pair.pathEnd2ChannelKey][MsgAcknowledgement],
+			SrcMsgTimeout:         pp.pathEnd2.messageCache.PacketFlow[pair.pathEnd2ChannelKey][MsgTimeout],
+			SrcMsgTimeoutOnClose:  pp.pathEnd2.messageCache.PacketFlow[pair.pathEnd2ChannelKey][MsgTimeoutOnClose],
+		}
+
+		pathEnd1ProcessRes[i] = pp.getUnrelayedPacketsAndAcksAndToDelete(ctx, pathEnd1PacketFlowMessages)
+		pathEnd2ProcessRes[i] = pp.getUnrelayedPacketsAndAcksAndToDelete(ctx, pathEnd2PacketFlowMessages)
 	}
 
-	retryCount := uint64(0)
+	// concatenate applicable messages for pathend
+	pathEnd1ConnectionMessages, pathEnd2ConnectionMessages := pp.connectionMessagesToSend(pathEnd1ConnectionHandshakeRes, pathEnd2ConnectionHandshakeRes)
+	pathEnd1ChannelMessages, pathEnd2ChannelMessages := pp.channelMessagesToSend(pathEnd1ChannelHandshakeRes, pathEnd2ChannelHandshakeRes)
+	pathEnd1PacketMessages, pathEnd2PacketMessages := pp.packetMessagesToSend(channelPairs, pathEnd1ProcessRes, pathEnd2ProcessRes)
 
-	if inProgress, ok := msgSendCache[message.connectionKey]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	msgSendCache[message.connectionKey] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
-	}
-}
-
-func (pathEnd *pathEndRuntime) trackSentChannelMessage(message channelIBCMessage) {
-	msgSendCache, ok := pathEnd.channelSendCache[message.action]
-	if !ok {
-		msgSendCache = make(channelKeySendCache)
-		pathEnd.channelSendCache[message.action] = msgSendCache
-	}
-
-	retryCount := uint64(0)
-
-	if inProgress, ok := msgSendCache[message.channelKey]; ok {
-		retryCount = inProgress.retryCount + 1
-	}
-
-	msgSendCache[message.channelKey] = inProgressSend{
-		sendHeight: pathEnd.latestBlock.Height,
-		message:    message.message,
-		retryCount: retryCount,
-	}
+	// now assemble and send messages in parallel
+	// if sending messages fails to one pathEnd, we don't need to halt sending to the other pathEnd.
+	var eg errgroup.Group
+	eg.Go(func() error {
+		if err := pp.assembleAndSendMessages(ctx, pp.pathEnd2, pp.pathEnd1, pathEnd1PacketMessages, pathEnd1ConnectionMessages, pathEnd1ChannelMessages); err != nil {
+			pp.log.Error("Error sending messages",
+				zap.String("src_chain_id", pp.pathEnd1.info.ChainID),
+				zap.String("dst_chain_id", pp.pathEnd2.info.ChainID),
+				zap.String("dst_client_id", pp.pathEnd2.info.ClientID),
+				zap.Error(err),
+			)
+			return err
+		}
+		return nil
+	})
+	eg.Go(func() error {
+		if err := pp.assembleAndSendMessages(ctx, pp.pathEnd1, pp.pathEnd2, pathEnd2PacketMessages, pathEnd2ConnectionMessages, pathEnd2ChannelMessages); err != nil {
+			pp.log.Error("Error sending messages",
+				zap.String("src_chain_id", pp.pathEnd2.info.ChainID),
+				zap.String("dst_chain_id", pp.pathEnd1.info.ChainID),
+				zap.String("dst_client_id", pp.pathEnd1.info.ClientID),
+				zap.Error(err),
+			)
+			return err
+		}
+		return nil
+	})
+	return eg.Wait()
 }
 
 func (pp *PathProcessor) assembleAndSendMessages(
